refactor: tidied up codebuild pipeline updater function
